refactor: move db outside the src folder
