+++ tsconfig.json
@@ -18,5 +18,5 @@
     "forceConsistentCasingInFileNames": false,
     "noFallthroughCasesInSwitch": false
   },
-  "include": [".eslintrc.js", "src", "types"]
+  "include": [".eslintrc.js", "src", "types", "db"]
 }
